moving the task logic to common GitVersionTasks
